Delete vendor and add it to .gitignore
Vendor files are useless when .git folders aren't there as well, because it disabled the ability to use composer update. Composer need the .git folder for dev versions which are cloned with git instead of using ZIPs by default. Composer.lock still ensures that there will be always the exact same dependencies.